move poc code into poc directory
