+++ server/app/routes/api/v1/poc/phone.test.js
@@ -1,9 +1,8 @@
-
 import { expect } from 'chai';
 
-import { ov_config } from '../../../lib/ov_config';
-import neo4j from '../../../lib/neo4j';
-import { appInit, base_uri, getObjs } from '../../../../test/lib/utils';
+import { appInit, base_uri, getObjs } from '../../../../../test/lib/utils';
+import { ov_config } from '../../../../lib/ov_config';
+import neo4j from '../../../../lib/neo4j';
 
 var api;
 var db;
